Extract native Computer runtime into workspace crate
