+++ crates/webcodex-runner/Cargo.toml
@@ -6,6 +6,7 @@ edition.workspace = true
 
 [dependencies]
 agent-client-protocol-schema = { version = "1.7.0", default-features = false }
+webcodex-computer = { path = "../webcodex-computer" }
 webcodex-runner-config = { path = "../webcodex-runner-config" }
 webcodex-core = { path = "../webcodex-core" }
 webcodex-persistent-shell = { path = "../webcodex-persistent-shell" }
@@ -38,40 +39,11 @@ webpki-roots = "0.26"
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
 
-[target.'cfg(any(target_os = "macos", target_os = "windows"))'.dependencies]
-image = { version = "0.25.9", default-features = false, features = ["jpeg"] }
-xcap = { version = "0.9.8", default-features = false }
-
-[target.'cfg(target_os = "macos")'.dependencies]
-block2 = "0.6.2"
-objc2-app-kit = { version = "0.3.2", default-features = false, features = ["NSPasteboard", "NSRunningApplication", "NSWorkspace", "block2"] }
-objc2-application-services = "0.3.2"
-objc2-core-foundation = "0.3.2"
-objc2-core-graphics = "0.3.2"
-objc2-foundation = { version = "0.3.2", default-features = false, features = ["NSArray", "NSBundle", "NSDictionary", "NSError", "NSFileManager", "NSObject", "NSPathUtilities", "NSString", "NSURL"] }
-
 [target.'cfg(windows)'.dependencies]
-windows = { version = "0.62.2", features = [
-    "Win32_Foundation",
-    "Win32_Graphics_Gdi",
-    "Win32_System_Com",
-    "Win32_System_DataExchange",
-    "Win32_System_Memory",
-    "Win32_System_Ole",
-    "Win32_System_Registry",
-    "Win32_UI_Accessibility",
-    "Win32_UI_Input_KeyboardAndMouse",
-    "Win32_UI_HiDpi",
-    "Win32_UI_Shell",
-    "Win32_UI_Shell_Common",
-    "Win32_UI_WindowsAndMessaging",
-] }
 windows-sys = { version = "0.61.2", features = [
     "Win32_Foundation",
     "Win32_Globalization",
-    "Win32_Graphics_Gdi",
     "Win32_Storage_FileSystem",
-    "Win32_Storage_Xps",
     "Win32_System_Threading",
 ] }
